+++ src/openapi.ts
@@ -21,7 +21,6 @@ import type {
 	ElysiaOpenAPIConfig,
 	MapJsonSchema
 } from './types'
-import { defineConfig } from 'tsup'
 
 export const capitalize = (word: string) =>
 	word.charAt(0).toUpperCase() + word.slice(1)